+++ retroauth/src/test/java/com/andretietz/retroauth/CredentialInterceptorTest.kt
@@ -1,13 +1,13 @@
 package com.andretietz.retroauth
 
 import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.anyOrNull
-import com.nhaarman.mockitokotlin2.doAnswer
-import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.never
-import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.doReturn
+import com.nhaarman.mockitokotlin2.times
+import com.nhaarman.mockitokotlin2.anyOrNull
+import com.nhaarman.mockitokotlin2.doAnswer
 import io.reactivex.Single
 import io.reactivex.observers.TestObserver
 import io.reactivex.schedulers.Schedulers
@@ -34,18 +34,13 @@ class CredentialInterceptorTest {
   @Test
   fun `unauthenticated call with successful response`() {
     serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn null as RequestType<String, String>?
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>>()
     val credentialStorage = mock<CredentialStorage<String, String, String>>()
     val authenticator = mock<Authenticator<String, String, String, String>>()
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -70,11 +65,6 @@ class CredentialInterceptorTest {
 
   @Test
   fun `authenticated call, no owner existing`() {
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
-
     // setup ownerstore, without any owner existing
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn null as String?
@@ -82,13 +72,15 @@ class CredentialInterceptorTest {
     }
 
     val credentialStorage = mock<CredentialStorage<String, String, String>>()
-    val authenticator = mock<Authenticator<String, String, String, String>>()
+    val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
+    }
 
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -102,7 +94,7 @@ class CredentialInterceptorTest {
       )
       .build().create(SomeApi::class.java)
 
-    val testObserver = api.someCall()
+    val testObserver = api.someAuthenticatedCall()
       .subscribeOn(Schedulers.io())
       .test()
 
@@ -115,10 +107,6 @@ class CredentialInterceptorTest {
   fun `authenticated call with successful response`() {
     serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
 
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -133,7 +121,10 @@ class CredentialInterceptorTest {
         override fun isCancelled() = false
       }
     }
+
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doReturn true
       on { authenticateRequest(any(), anyString()) } doAnswer { invocationOnMock ->
         (invocationOnMock.arguments[0] as Request)
@@ -146,8 +137,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -161,7 +151,7 @@ class CredentialInterceptorTest {
       )
       .build().create(SomeApi::class.java)
 
-    val testObserver = api.someCall()
+    val testObserver = api.someAuthenticatedCall()
       .subscribeOn(Schedulers.io())
       .test()
 
@@ -175,10 +165,6 @@ class CredentialInterceptorTest {
   fun `Invalid token, refreshes token`() {
     serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
 
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -194,6 +180,8 @@ class CredentialInterceptorTest {
       }
     }
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doReturn false // token is invalid
       on { authenticateRequest(any(), anyString()) } doAnswer { invocationOnMock ->
         (invocationOnMock.arguments[0] as Request)
@@ -209,8 +197,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -224,7 +211,7 @@ class CredentialInterceptorTest {
       )
       .build().create(SomeApi::class.java)
 
-    val testObserver = api.someCall()
+    val testObserver = api.someAuthenticatedCall()
       .subscribeOn(Schedulers.io())
       .test()
 
@@ -243,10 +230,6 @@ class CredentialInterceptorTest {
   fun `Invalid token, refresh token fails`() {
     serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
 
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -262,6 +245,8 @@ class CredentialInterceptorTest {
       }
     }
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doReturn false // token is invalid
       on { authenticateRequest(any(), anyString()) } doAnswer { invocationOnMock ->
         (invocationOnMock.arguments[0] as Request)
@@ -277,8 +262,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -292,7 +276,7 @@ class CredentialInterceptorTest {
       )
       .build().create(SomeApi::class.java)
 
-    val testObserver = api.someCall()
+    val testObserver = api.someAuthenticatedCall()
       .subscribeOn(Schedulers.io())
       .test()
 
@@ -313,10 +297,6 @@ class CredentialInterceptorTest {
     serverRule.server.enqueue(MockResponse().setResponseCode(401))
     serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
 
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -332,6 +312,8 @@ class CredentialInterceptorTest {
       }
     }
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doReturn true
       on { authenticateRequest(any(), anyString()) } doAnswer { invocationOnMock ->
         (invocationOnMock.arguments[0] as Request)
@@ -352,8 +334,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -367,7 +348,7 @@ class CredentialInterceptorTest {
       )
       .build().create(SomeApi::class.java)
 
-    val testObserver = api.someCall()
+    val testObserver = api.someAuthenticatedCall()
       .subscribeOn(Schedulers.io())
       .test()
 
@@ -380,10 +361,6 @@ class CredentialInterceptorTest {
   @Test
   fun `Invalid token, refreshes token, 200 calls`() {
     val range = IntRange(0, 199)
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -400,6 +377,8 @@ class CredentialInterceptorTest {
       }
     }
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doAnswer { invocationOnMock ->
         invocationOnMock.arguments[0] as String == "credential"
       }
@@ -421,8 +400,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -439,9 +417,11 @@ class CredentialInterceptorTest {
     val calls = mutableListOf<TestObserver<Data>>()
     for (i in range) {
       serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
-      calls.add(api.someCall()
-        .subscribeOn(Schedulers.io()) // makes sure a new thread is created foreach call
-        .test())
+      calls.add(
+        api.someAuthenticatedCall()
+          .subscribeOn(Schedulers.io()) // makes sure a new thread is created foreach call
+          .test()
+      )
     }
 
     for (i in range) {
@@ -460,10 +440,7 @@ class CredentialInterceptorTest {
   @Test
   fun `Invalid token, refreshes token an error occurs, 200 calls`() {
     val range = IntRange(0, 199)
-    val methodCache = mock<MethodCache<String, String>> {
-      // when this returns null, the call is not recognized as authenticated call.
-      on { getCredentialType(any()) } doReturn RequestType("credentialType", "ownerType")
-    }
+
     val ownerStorage = mock<OwnerStorage<String, String, String>> {
       on { getActiveOwner(anyString()) } doReturn "owner"
     }
@@ -479,6 +456,8 @@ class CredentialInterceptorTest {
       }
     }
     val authenticator = mock<Authenticator<String, String, String, String>> {
+      on { getCredentialType(any()) } doReturn CREDENTIAL_TYPE
+      on { getOwnerType(any()) } doReturn OWNER_TYPE
       on { isCredentialValid(anyString()) } doReturn false
       on {
         refreshCredentials(anyString(), anyString(), anyString())
@@ -491,8 +470,7 @@ class CredentialInterceptorTest {
     val interceptor = CredentialInterceptor(
       authenticator,
       ownerStorage,
-      credentialStorage,
-      methodCache
+      credentialStorage
     )
 
     val api = Retrofit.Builder()
@@ -509,9 +487,11 @@ class CredentialInterceptorTest {
     val calls = mutableListOf<TestObserver<Data>>()
     for (i in range) {
       serverRule.server.enqueue(MockResponse().setBody("{\"data\": \"testdata\"}"))
-      calls.add(api.someCall()
-        .subscribeOn(Schedulers.io()) // makes sure a new thread is created foreach call
-        .test())
+      calls.add(
+        api.someAuthenticatedCall()
+          .subscribeOn(Schedulers.io()) // makes sure a new thread is created foreach call
+          .test()
+      )
     }
     for (i in range) {
       calls[i].awaitTerminalEvent()
@@ -526,6 +506,11 @@ class CredentialInterceptorTest {
     // authenticate request -> never
     verify(authenticator, never()).authenticateRequest(any(), anyString())
   }
+
+  companion object {
+    private const val OWNER_TYPE = "owner_type"
+    private const val CREDENTIAL_TYPE = "credential_type"
+  }
 }
 
 data class Data(val data: String)
@@ -533,4 +518,8 @@ data class Data(val data: String)
 interface SomeApi {
   @GET("/some/path")
   fun someCall(): Single<Data>
+
+  @Authenticated
+  @GET("/some/other/path")
+  fun someAuthenticatedCall(): Single<Data>
 }